refactor: separate modal and header
